fix(backend): moved loggers file inside app folder to resolve module discovery in docker image
